Console: Split off a few classes
